backend/requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
psycopg2-binary>=2.9
python-dotenv>=1.0
gunicorn>=20.1
whitenoise>=6.5
# Optional useful tools
django-cors-headers>=3.13
Pillow>=10.0
Pillow>=10.0
pypdf>=4.0.0
